requirements.txt
pytest==7.1.2
requests==2.27.1
vcrpy==4.1.1
